// src/lib/api/generated/model/publicBrandingAssetMetadataSchema.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The canonical backend API for COMPASS clients.
 * OpenAPI spec version: 1.0.0
 */

export interface PublicBrandingAssetMetadataSchema {
  alt_text: string;
  asset_type: string;
  content_type: string;
  delivery_mode: string;
  delivery_path: string;
  display_order: number;
  effective_from?: string | null;
  effective_until?: string | null;
  height: number;
  id: string;
  placement: string;
  width: number;
}