// src/lib/api/generated/model/publicLinkItemSchema.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The canonical backend API for COMPASS clients.
 * OpenAPI spec version: 1.0.0
 */

export interface PublicLinkItemSchema {
  display_order: number;
  label: string;
  link_type: string;
  owner_label: string;
  owner_type: string;
  placement: string;
  url: string;
}